updated to visual studio 2026 copied code for drawing different types of snow flakes from https://github.com/riyasy/RainProject/pull/11

+++ Src/let-it-rain/DisplayWindow.cpp
@@ -620,7 +620,7 @@ void DisplayWindow::UpdateRainDrops()
 
 void DisplayWindow::UpdateSnowFlakes()
 {
-	const int noOfFlakesToGenerate = GeneralSettings.MaxParticles * 20 - SnowFlakes.size();
+	const int noOfFlakesToGenerate = GeneralSettings.MaxParticles * 20 - static_cast<int>(SnowFlakes.size());
 
 	if (noOfFlakesToGenerate > 0)
 	{
